requirements.txt
fastapi
uvicorn
PyJWT
passlib
SQLAlchemy
pydantic
argon2-cffi
pydantic_settings
python-multipart